+++ Sources/VComponents/Components/Modals/Modal/VModalUIModel.swift
@@ -13,7 +13,14 @@ import VCore
 @available(watchOS, unavailable)
 public struct VModalUIModel: Sendable {
     // MARK: Properties - Global
-    var presentationHostSubUIModel: PresentationHostUIModel { .init() }
+    var presentationHostSubUIModel: PresentationHostUIModel {
+        var uiModel: PresentationHostUIModel = .init()
+        uiModel.preferredDimmingViewColor = preferredDimmingViewColor
+        return uiModel
+    }
+    
+    /// Preferred dimming color, that overrides a shared color from Presentation Host layer, when only this modal is presented.
+    public var preferredDimmingViewColor: Color?
     
     /// Modal size group.
     /// Set to `fixed` with `(0.9, 0.6)` `fraction`s in portrait and `fixed` with `(0.5, 1)` `fraction`s in landscape on `iOS`.